Generalize Skylark parser for Bazel and Buck #1678
    * Update tests
